test: fold former known-limitation tests into DatetimeFormatTests
